json test fixes; s3 special chrs; spec tweak; empty streams; state refact

+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/data/AirbyteValue.kt
@@ -22,7 +22,7 @@ sealed interface AirbyteValue {
                 null -> NullValue
                 is String -> StringValue(value)
                 is Boolean -> BooleanValue(value)
-                is Int -> IntValue(value)
+                is Int -> IntegerValue(value.toLong())
                 is Long -> IntegerValue(value)
                 is BigInteger -> IntegerValue(value)
                 is Double -> NumberValue(BigDecimal.valueOf(value))
